Fix for reproducability of results
Encountered reproducability issues when running more complex problems against Jsprit.
This issue manifests when the ruinedJobSet is created in RuinAndRecreateModule.
Ordering in the HashSet is not guaranteed. When re-running the same problem occasionally
the order in the HashSet of some Jobs were slightly different. When jobs are shuffled inside
an InsertionStrategy (e.g. BestInsertion and BestInsertionConcurrent) the different ordering
is exhaserbated. The `Collections.sort(unassignedJobList, new AccordingToPriorities());`
ensures that the order of equal items is guaranteed, resulting in different insertions for
different runs of the same problem.

This issue only showed itself when running with threads. The reason for this is unkown.
Notably, when disabling clustersRegret and stringRegret the issue no longer occurred.
The reason for this is unclear but is likely related to the order of items in the unassignedJobs list
returned from the regret classes.

From tests, simply swapping to a LinkedHashSet did not remove this ordering problem so the
unassignedJobs were inserted to a list and sorted. Since an ArrayList is faster on iteration
the ArrayList was kept, however re-inserting the sortedUnassignedJobs may be preferrable.