Group activities together if possible by adjusting their times
After all activities have been inserted in a route and the ActivityTimeTracker has
updated the times of the activites, check if some of them could be group together.

A activity belongs to a group of activities if their locations are the same and if they
could be served at the same time.
This is achieved by iterating over the activites for a route when
`UpdateActivityTimes.finish` is called.